+++ kinto_http/tests/test_logging.py
@@ -13,32 +13,32 @@ class BucketLoggingTest(unittest.TestCase):
         mock_response(self.session)
 
     def test_create_bucket_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_bucket(id="buck", data={"foo": "bar"})
             mocked_logger.info.assert_called_with("Create bucket 'buck'")
 
     def test_update_bucket_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.update_bucket(id="buck", data={"foo": "bar"})
             mocked_logger.info.assert_called_with("Update bucket 'buck'")
 
     def test_patch_bucket_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.patch_bucket(id="buck", data={"foo": "bar"})
             mocked_logger.info.assert_called_with("Patch bucket 'buck'")
 
     def test_get_bucket_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.get_bucket(id="buck")
             mocked_logger.info.assert_called_with("Get bucket 'buck'")
 
     def test_delete_bucket_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.delete_bucket(id="buck")
             mocked_logger.info.assert_called_with("Delete bucket 'buck'")
 
     def test_delete_buckets_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.delete_buckets()
             mocked_logger.info.assert_called_with("Delete buckets")
 
@@ -50,14 +50,14 @@ class GroupLoggingTest(unittest.TestCase):
         mock_response(self.session)
 
     def test_create_group_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_group(
                 id="mozilla", bucket="buck", data={"foo": "bar"}, permissions={"write": ["blah"]}
             )
             mocked_logger.info.assert_called_with("Create group 'mozilla' in bucket 'buck'")
 
     def test_update_group_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.update_group(
                 data={"foo": "bar"},
                 id="mozilla",
@@ -67,7 +67,7 @@ class GroupLoggingTest(unittest.TestCase):
             mocked_logger.info.assert_called_with("Update group 'mozilla' in bucket 'buck'")
 
     def test_patch_group_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.patch_group(
                 data={"foo": "bar"},
                 id="mozilla",
@@ -77,17 +77,17 @@ class GroupLoggingTest(unittest.TestCase):
             mocked_logger.info.assert_called_with("Patch group 'mozilla' in bucket 'buck'")
 
     def test_get_group_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.get_group(id="mozilla", bucket="buck")
             mocked_logger.info.assert_called_with("Get group 'mozilla' in bucket 'buck'")
 
     def test_delete_group_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.delete_group(id="mozilla", bucket="buck")
             mocked_logger.info.assert_called_with("Delete group 'mozilla' in bucket 'buck'")
 
     def test_delete_groups_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.delete_groups(bucket="buck")
             mocked_logger.info.assert_called_with("Delete groups in bucket 'buck'")
 
@@ -99,14 +99,14 @@ class CollectionLoggingTest(unittest.TestCase):
         mock_response(self.session)
 
     def test_create_collection_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_collection(
                 id="mozilla", bucket="buck", data={"foo": "bar"}, permissions={"write": ["blah"]}
             )
             mocked_logger.info.assert_called_with("Create collection 'mozilla' in bucket 'buck'")
 
     def test_update_collection_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.update_collection(
                 data={"foo": "bar"},
                 id="mozilla",
@@ -116,7 +116,7 @@ class CollectionLoggingTest(unittest.TestCase):
             mocked_logger.info.assert_called_with("Update collection 'mozilla' in bucket 'buck'")
 
     def test_patch_collection_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.patch_collection(
                 data={"foo": "bar"},
                 id="mozilla",
@@ -126,17 +126,17 @@ class CollectionLoggingTest(unittest.TestCase):
             mocked_logger.info.assert_called_with("Patch collection 'mozilla' in bucket 'buck'")
 
     def test_get_collection_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.get_collection(id="mozilla", bucket="buck")
             mocked_logger.info.assert_called_with("Get collection 'mozilla' in bucket 'buck'")
 
     def test_delete_collection_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.delete_collection(id="mozilla", bucket="buck")
             mocked_logger.info.assert_called_with("Delete collection 'mozilla' in bucket 'buck'")
 
     def test_delete_collections_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.delete_collections(bucket="buck")
             mocked_logger.info.assert_called_with("Delete collections in bucket 'buck'")
 
@@ -148,7 +148,7 @@ class RecordLoggingTest(unittest.TestCase):
         mock_response(self.session)
 
     def test_create_record_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_bucket(id="buck")
             self.client.create_collection(id="mozilla", bucket="buck")
             self.client.create_record(
@@ -163,7 +163,7 @@ class RecordLoggingTest(unittest.TestCase):
             )
 
     def test_update_record_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_bucket(id="buck")
             self.client.create_collection(bucket="buck", id="mozilla")
             self.client.update_record(
@@ -174,7 +174,7 @@ class RecordLoggingTest(unittest.TestCase):
             )
 
     def test_patch_record_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_bucket(id="buck")
             self.client.create_collection(bucket="buck", id="mozilla")
             self.client.patch_record(
@@ -185,7 +185,7 @@ class RecordLoggingTest(unittest.TestCase):
             )
 
     def test_get_record_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_bucket(id="buck")
             self.client.create_collection(id="mozilla", bucket="buck")
             self.client.get_record(id="fake-record", bucket="buck", collection="mozilla")
@@ -194,7 +194,7 @@ class RecordLoggingTest(unittest.TestCase):
             )
 
     def test_delete_record_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_bucket(id="buck")
             self.client.create_collection(id="mozilla", bucket="buck")
             self.client.delete_record(id="fake-record", bucket="buck", collection="mozilla")
@@ -203,7 +203,7 @@ class RecordLoggingTest(unittest.TestCase):
             )
 
     def test_delete_records_logs_info_message(self):
-        with mock.patch("kinto_http.logger") as mocked_logger:
+        with mock.patch("kinto_http.client.logger") as mocked_logger:
             self.client.create_bucket(id="buck")
             self.client.create_collection(id="mozilla", bucket="buck")
             self.client.delete_records(bucket="buck", collection="mozilla")